requirements3.txt
aiohttp==1.3.3
appdirs==1.4.1
async-timeout==1.1.0
chardet==2.3.0
feedparser==5.2.1
multidict==2.1.4
packaging==16.8
pyparsing==2.1.10
PyYAML==3.12
six==1.10.0
telepot==10.4
urllib3==1.20
yarl==0.9.8
